tweaked formatting & optimized errorHandler
deck creation now checks if a deck already exists
fixed double-assigning size variable
cardHandler is now able to delete a deck of cards
cardHandler now always returns with a success boolean & an optional reason

+++ handlers/cardHandler.js
@@ -1,19 +1,29 @@
 const { EmbedBuilder }  = require("discord.js")
 const jsonfile	        = require("jsonfile")
 const fs                = require("fs")
+
 const deck_template     = [ "Ace", "Ace", "Ace", "Ace", 2, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 4, 5, 5, 5, 5, 6, 6, 6, 6, 7, 7, 7, 7, 8, 8, 8, 8, 9, 9, 9, 9, 10, 10, 10, 10, "Jack", "Jack", "Jack", "Jack", "Queen", "Queen", "Queen", "Queen", "King", "King", "King", "King"]
 
 //create a new deck of cards 
 async function create(UID, size)
 {
-    if(size)    size = size
-    else        size = 1
+    if(fs.existsSync(`database/deck_${UID}`)) return{ success: false, reason: "Deck already exists"}
 
-    const deck = Array(size).fill(deck_template).flat()
+    try
+    {
+        size = size || 1
 
-    console.log(size, "\n", deck)
+        const deck = Array(size).fill(deck_template).flat()
 
-    jsonfile.writeFileSync(`database/deck_${UID}`, deck)
+        jsonfile.writeFileSync(`database/deck_${UID}`, deck)
+
+        return{ success: true }
+    }
+    catch(err)
+    {
+        console.log(err)
+        return{ success: false, reason: "Failed to create new deck"}
+    }
 }
 
 //draw from an existing deck of cards
@@ -29,9 +39,26 @@ async function burn(ID, amount)
 }
 
 //delete a deck 
-async function remove(ID, amount)
+async function remove(ID)
 {
-    
+    try { jsonfile.readFileSync(`database/deck_${ID}`) }
+    catch   
+    { 
+        reason = "Couldn't find deck"
+        return{ success: false, reason: "Deck not found" }
+    }
+
+
+    await fs.unlink(`database/deck_${ID}`, (err)=> 
+    {
+        if(err)
+        {
+            console.log(err)
+            return{ success: false, reason: "Unable to delete deck, error has been logged."}
+        }
+    })
+
+    return{ success: true }
 }
